+++ R/problems.R
@@ -32,7 +32,7 @@ problems <- function(x = .Last.value, lazy = FALSE) {
   if (typeof(probs) != "externalptr") {
     cli::cli_abort(c(
       "The {.arg x} argument of {.fun vroom::problems} must be a data frame created by vroom:",
-      x = "{.arg x} seems to have been created with something else, maybe readr?"
+      x = "{.arg x} seems to have been created with something else, maybe {.pkg readr}?"
     ))
   }
   probs <- vroom_errors_(probs)