+++ NEWS.md
@@ -16,6 +16,9 @@
 
 * Columns specified as having type "number" (requested via `col_number()` or `"number"` or `'n'`) or "skip" (requested via `col_skip()` or `"skip"` or `_` or `-`) now work in the case where 0 rows of data are parsed (#427, #540, #548).
 
+* vroom takes the recommended approach for phasing out usage of the non-API
+  entry points `SETLENGTH`, `SET_GROWABLE_BIT`, and `SET_TRUELENGTH` (#582).
+
 # vroom 1.6.7
 
 * `locale(encoding =)` now warns, instead of errors, when the encoding cannot be found in `iconvlist()` return value. This removes an unnecessary blocker on platforms like Alpine Linux where the output doesn't reflect actual capabilities.